refactor: legacy 삭제 및 패키지이동
