--- app/src/main/java/com/techcourse/controller/AnnotationLoginController.java
@@ -1,56 +1,0 @@
-package com.techcourse.controller;
-
-import com.techcourse.domain.User;
-import com.techcourse.repository.InMemoryUserRepository;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
-import nextstep.mvc.view.JspView;
-import nextstep.mvc.view.ModelAndView;
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.support.RequestMethod;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import static nextstep.mvc.view.JspView.*;
-
-@Controller
-public class AnnotationLoginController {
-
-    private static final Logger log = LoggerFactory.getLogger(AnnotationLoginController.class);
-
-    @RequestMapping(value = "/login", method = RequestMethod.POST)
-    public ModelAndView login(HttpServletRequest request, HttpServletResponse response) {
-        if (UserSession.isLoggedIn(request.getSession())) {
-            return new ModelAndView(new JspView(REDIRECT_PREFIX + "/index.jsp"));
-        }
-
-        return InMemoryUserRepository.findByAccount(request.getParameter("account"))
-                .map(user -> {
-                    log.info("User : {}", user);
-                    return login(request, user);
-                })
-                .orElse(new ModelAndView(new JspView(REDIRECT_PREFIX + "/401.jsp")));
-    }
-
-    @RequestMapping(value = "/login", method = RequestMethod.GET)
-    public ModelAndView loginView(HttpServletRequest request, HttpServletResponse response) {
-        return UserSession.getUserFrom(request.getSession())
-                .map(user -> {
-                    log.info("logged in {}", user.getAccount());
-                    return new ModelAndView(new JspView(REDIRECT_PREFIX + "/index.jsp"));
-                })
-                .orElse(new ModelAndView(new JspView("/login.jsp")));
-    }
-
-    public ModelAndView login(HttpServletRequest request, User user) {
-        if (user.checkPassword(request.getParameter("password"))) {
-            final HttpSession session = request.getSession();
-            session.setAttribute(UserSession.SESSION_KEY, user);
-            return new ModelAndView(new JspView(REDIRECT_PREFIX + "/index.jsp"));
-        } else {
-            return new ModelAndView(new JspView(REDIRECT_PREFIX + "/401.jsp"));
-        }
-    }
-}
