--- app/src/main/java/com/techcourse/controller/AnnotationRegisterController.java
@@ -1,33 +1,0 @@
-package com.techcourse.controller;
-
-import com.techcourse.domain.User;
-import com.techcourse.repository.InMemoryUserRepository;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.view.JspView;
-import nextstep.mvc.view.ModelAndView;
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.support.RequestMethod;
-
-import static nextstep.mvc.view.JspView.*;
-
-@Controller
-public class AnnotationRegisterController {
-
-    @RequestMapping(value = "/register", method = RequestMethod.POST)
-    public ModelAndView register(HttpServletRequest request, HttpServletResponse response) {
-        final User user = new User(
-                request.getParameter("account"),
-                request.getParameter("password"),
-                request.getParameter("email"));
-        InMemoryUserRepository.save(user);
-
-        return new ModelAndView(new JspView(REDIRECT_PREFIX + "/index.jsp"));
-    }
-
-    @RequestMapping(value = "/register", method = RequestMethod.GET)
-    public ModelAndView registerView(HttpServletRequest request, HttpServletResponse response) {
-        return new ModelAndView(new JspView("/register.jsp"));
-    }
-}
