+++ app/src/main/java/com/techcourse/controller/ForwardController.java
@@ -9,7 +9,7 @@ import nextstep.web.annotation.RequestMapping;
 import nextstep.web.support.RequestMethod;
 
 @Controller
-public class AnnotationForwardController {
+public class ForwardController {
 
     @RequestMapping(value = "/", method = RequestMethod.GET)
     public ModelAndView root(HttpServletRequest request, HttpServletResponse response) {
