+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -5,35 +5,52 @@ import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import nextstep.mvc.controller.asis.Controller;
+import nextstep.mvc.view.JspView;
+import nextstep.mvc.view.ModelAndView;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.support.RequestMethod;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class LoginController implements Controller {
+import static nextstep.mvc.view.JspView.*;
+
+@Controller
+public class LoginController {
 
     private static final Logger log = LoggerFactory.getLogger(LoginController.class);
 
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse res) throws Exception {
-        if (UserSession.isLoggedIn(req.getSession())) {
-            return "redirect:/index.jsp";
+    @RequestMapping(value = "/login", method = RequestMethod.POST)
+    public ModelAndView login(HttpServletRequest request, HttpServletResponse response) {
+        if (UserSession.isLoggedIn(request.getSession())) {
+            return new ModelAndView(new JspView(REDIRECT_PREFIX + "/index.jsp"));
         }
 
-        return InMemoryUserRepository.findByAccount(req.getParameter("account"))
+        return InMemoryUserRepository.findByAccount(request.getParameter("account"))
                 .map(user -> {
                     log.info("User : {}", user);
-                    return login(req, user);
+                    return login(request, user);
                 })
-                .orElse("redirect:/401.jsp");
+                .orElse(new ModelAndView(new JspView(REDIRECT_PREFIX + "/401.jsp")));
     }
 
-    private String login(HttpServletRequest request, User user) {
+    @RequestMapping(value = "/login", method = RequestMethod.GET)
+    public ModelAndView loginView(HttpServletRequest request, HttpServletResponse response) {
+        return UserSession.getUserFrom(request.getSession())
+                .map(user -> {
+                    log.info("logged in {}", user.getAccount());
+                    return new ModelAndView(new JspView(REDIRECT_PREFIX + "/index.jsp"));
+                })
+                .orElse(new ModelAndView(new JspView("/login.jsp")));
+    }
+
+    public ModelAndView login(HttpServletRequest request, User user) {
         if (user.checkPassword(request.getParameter("password"))) {
             final HttpSession session = request.getSession();
             session.setAttribute(UserSession.SESSION_KEY, user);
-            return "redirect:/index.jsp";
+            return new ModelAndView(new JspView(REDIRECT_PREFIX + "/index.jsp"));
         } else {
-            return "redirect:/401.jsp";
+            return new ModelAndView(new JspView(REDIRECT_PREFIX + "/401.jsp"));
         }
     }
 }
