+++ app/src/main/java/com/techcourse/controller/LogoutController.java
@@ -3,14 +3,21 @@ package com.techcourse.controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import nextstep.mvc.controller.asis.Controller;
+import nextstep.mvc.view.JspView;
+import nextstep.mvc.view.ModelAndView;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.support.RequestMethod;
 
-public class LogoutController implements Controller {
+import static nextstep.mvc.view.JspView.*;
 
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse res) throws Exception {
-        final HttpSession session = req.getSession();
+@Controller
+public class LogoutController {
+
+    @RequestMapping(value = "/logout", method = RequestMethod.GET)
+    public ModelAndView logout(HttpServletRequest request, HttpServletResponse response) {
+        final HttpSession session = request.getSession();
         session.removeAttribute(UserSession.SESSION_KEY);
-        return "redirect:/";
+        return new ModelAndView(new JspView(REDIRECT_PREFIX + "/login.jsp"));
     }
 }
