+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -5,9 +5,9 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.tobe.HandlerAdapters;
-import nextstep.mvc.controller.tobe.HandlerExecution;
-import nextstep.mvc.controller.tobe.HandlerMappings;
+import nextstep.mvc.controller.HandlerAdapters;
+import nextstep.mvc.controller.HandlerExecution;
+import nextstep.mvc.controller.HandlerMappings;
 import nextstep.mvc.exception.AdapterNotFoundException;
 import nextstep.mvc.exception.HandlerNotFoundException;
 import nextstep.mvc.view.ModelAndView;
