+++ mvc/src/main/java/nextstep/mvc/controller/ExecutionHandlerAdapter.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
