+++ mvc/src/main/java/nextstep/mvc/controller/HandlerAdapters.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.controller;
 
 import nextstep.mvc.HandlerAdapter;
 import nextstep.mvc.exception.AdapterNotFoundException;
