+++ mvc/src/main/java/nextstep/mvc/controller/HandlerMappings.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import nextstep.mvc.HandlerMapping;
@@ -7,7 +7,6 @@ import nextstep.mvc.exception.HandlerNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 
 public class HandlerMappings {
