--- mvc/src/main/java/nextstep/mvc/controller/tobe/ControllerInterfaceHandlerAdapter.java
@@ -1,24 +1,0 @@
-package nextstep.mvc.controller.tobe;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.HandlerAdapter;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.view.JspView;
-import nextstep.mvc.view.ModelAndView;
-
-public class ControllerInterfaceHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        Controller handlerObject = (Controller) handler;
-        return new ModelAndView(
-                new JspView(handlerObject.execute(request, response))
-        );
-    }
-}
